restaurant_management/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.2
Django==5.2.4
django-filter==25.1
djangorestframework==3.16.0
dnspython==2.8.0
email-validator==2.3.0
idna==3.10
pillow==11.3.0
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
